# openapi.yaml
openapi: 3.0.3
info:
  title: Chat Backend API Gateway
  version: 0.1.0
  description: Publiczne REST API gateway dla backendu czatu tekstowego. Specyfikacja opisuje wyłącznie endpointy publiczne, bez endpointów wewnętrznych między mikroserwisami.
servers:
  - url: http://localhost:3000
    description: Local Docker Compose API Gateway
...
paths:
...
  /users:
    post:
      tags: [Users]
      summary: Create user
      description: Tworzy użytkownika w PostgreSQL przez pg-service. Email musi być unikalny.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            examples:
              createUser:
                value:
                  email: alice@example.test
                  username: alice
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              examples:
                created:
                  value:
                    id: 5f5226a3-0988-40ca-806e-7a86c995f597
                    email: alice@example.test
                    username: alice
                    createdAt: "2026-05-07T10:00:00.000Z"
                    deletedAt: null
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user
      description: Zwraca użytkownika po identyfikatorze UUID.
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              examples:
                found:
                  value:
                    id: 5f5226a3-0988-40ca-806e-7a86c995f597
                    email: alice@example.test
                    username: alice
                    createdAt: "2026-05-07T10:00:00.000Z"
                    deletedAt: null
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Users]
      summary: Soft delete user
      description: Oznacza użytkownika jako usuniętego przez deletedAt. Historia wiadomości i członkostwa pozostają zachowane.
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: User soft deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              examples:
                deleted:
                  value:
                    id: 5f5226a3-0988-40ca-806e-7a86c995f597
                    email: alice@example.test
                    username: alice
                    createdAt: "2026-05-07T10:00:00.000Z"
                    deletedAt: "2026-05-07T11:00:00.000Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /users/{userId}/conversations:
    get:
      tags: [Conversations]
      summary: List user conversations
      description: Zwraca konwersacje użytkownika posortowane po lastMessageAt DESC, potem createdAt DESC.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 5f5226a3-0988-40ca-806e-7a86c995f597
      responses:
        "200":
          description: User conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserConversation"
              examples:
                success:
                  value:
                    - id: 0d38e5ec-33f4-4d23-90b6-210f6f64391a
                      type: group
                      title: Study group
                      role: admin
                      joinedAt: "2026-05-07T10:00:00.000Z"
                      lastMessageAt: "2026-05-07T10:30:00.000Z"
                      lastSeq: 12
                      createdAt: "2026-05-07T10:00:00.000Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /conversations/{conversationId}/messages:
    post:
      tags: [Messages]
      summary: Send message
      description: Zapisuje wiadomość w operacji hybrydowej PG + Mongo. Gateway sprawdza członkostwo autora, rezerwuje seq w PostgreSQL, zapisuje treść w MongoDB i finalizuje message pointer w PostgreSQL. Przy błędzie finalizacji wykonuje kompensację MongoDB.
      parameters:
        - $ref: "#/components/parameters/ConversationIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
            examples:
              textMessage:
                value:
                  authorId: 5f5226a3-0988-40ca-806e-7a86c995f597
                  body: Hello from gateway
                  attachments: []
                  clientMessageId: mobile-123
              attachmentOnly:
                value:
                  authorId: 5f5226a3-0988-40ca-806e-7a86c995f597
                  body: ""
                  attachments:
                    - fileName: note.txt
                      mimeType: text/plain
                      size: 128
                      storageKey: attachments/note.txt
      responses:
        "201":
          description: Message stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              examples:
                created:
                  value:
                    _id: 663d11223b9d5d6cc4c2f111
                    conversationId: 0d38e5ec-33f4-4d23-90b6-210f6f64391a
                    authorId: 5f5226a3-0988-40ca-806e-7a86c995f597
                    seq: 1
                    body: Hello from gateway
                    createdAt: "2026-05-07T10:30:00.000Z"
                    editedAt: null
                    deliveryStatus: server_received
                    attachments: []
                    clientMessageId: mobile-123
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags: [Messages]
      summary: List conversation messages
      description: Wymaga requesterId w query albo nagłówka X-User-Id. Gateway sprawdza członkostwo w pg-service, a mongo-service pobiera wiadomości przez indeks conversationId + seq. Wynik jest sortowany chronologicznie po seq ASC.
      parameters:
        - $ref: "#/components/parameters/ConversationIdPath"
        - $ref: "#/components/parameters/RequesterIdQuery"
        - $ref: "#/components/parameters/XUserIdHeader"
        - name: afterSeq
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          example: 10
        - name: beforeSeq
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          example: 30
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: mimeType
          in: query
          required: false
          schema:
            type: string
          example: image/png,text/plain
      responses:
        "200":
          description: Messages sorted by seq ASC
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
              examples:
                success:
                  value:
                    - _id: 663d11223b9d5d6cc4c2f111
                      conversationId: 0d38e5ec-33f4-4d23-90b6-210f6f64391a
                      authorId: 5f5226a3-0988-40ca-806e-7a86c995f597
                      seq: 12
                      body: Hello
                      createdAt: "2026-05-07T10:30:00.000Z"
                      editedAt: null
                      deliveryStatus: server_received
                      attachments: []
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    UserIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 5f5226a3-0988-40ca-806e-7a86c995f597
    ConversationIdPath:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 0d38e5ec-33f4-4d23-90b6-210f6f64391a
    RequesterIdQuery:
      name: requesterId
      in: query
      required: false
      description: UUID użytkownika wykonującego odczyt. Alternatywnie można podać X-User-Id.
      schema:
        type: string
        format: uuid
      example: 5f5226a3-0988-40ca-806e-7a86c995f597
    XUserIdHeader:
      name: X-User-Id
      in: header
      required: false
      description: UUID użytkownika wykonującego odczyt, alternatywa dla requesterId.
      schema:
        type: string
        format: uuid
      example: 5f5226a3-0988-40ca-806e-7a86c995f597
...
  schemas:
...
    User:
      type: object
      required: [id, email, username, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    UserConversation:
      type: object
      required: [id, type, role, joinedAt, lastSeq, createdAt]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [direct, group]
        title:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, member]
        joinedAt:
          type: string
          format: date-time
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        lastSeq:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Message:
      type: object
      required: [_id, conversationId, authorId, seq, body, createdAt, editedAt, deliveryStatus, attachments]
      properties:
        _id:
          type: string
          example: 663d11223b9d5d6cc4c2f111
        conversationId:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
        seq:
          type: integer
          minimum: 1
        body:
          type: string
          maxLength: 2000
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
          nullable: true
        deliveryStatus:
          type: string
          enum: [server_received, delivered, read]
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        clientMessageId:
          type: string
        score:
          type: number
          description: Obecne tylko w wynikach wyszukiwania tekstowego.
    Attachment:
      type: object
      required: [fileName, mimeType, size, storageKey]
      properties:
        fileName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
          minimum: 0
        storageKey:
          type: string
...
    CreateUserRequest:
      type: object
      required: [email, username]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 1
...
    SendMessageRequest:
      type: object
      required: [authorId, body]
      properties:
        authorId:
          type: string
          format: uuid
        body:
          type: string
          maxLength: 2000
          description: Może być puste tylko wtedy, gdy attachments nie jest puste.
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
          default: []
        clientMessageId:
          type: string
          description: Opcjonalny klucz idempotencji w krótkim oknie.
...
    ErrorResponse:
      type: object
      required: [error, code, details]
      properties:
        error:
          type: string
          example: User is not a conversation member
        code:
          type: string
          example: NOT_MEMBER
        details:
          type: object
          nullable: true
          additionalProperties: true
          example:
            conversationId: 0d38e5ec-33f4-4d23-90b6-210f6f64391a
  responses:
    BadRequest:
      description: Validation or malformed request error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation:
              value:
                error: Field is required
                code: VALIDATION_ERROR
                details:
                  field: requesterId
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            forbidden:
              value:
                error: User is not a conversation member
                code: NOT_MEMBER
                details: null
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error: Resource not found
                code: NOT_FOUND
                details: null
    Conflict:
      description: Conflict, duplicate resource or idempotency conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            conflict:
              value:
                error: Unique constraint violation
                code: PRISMA_UNIQUE_CONSTRAINT
                details:
                  target:
                    - email
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internal:
              value:
                error: Internal server error
                code: INTERNAL_SERVER_ERROR
                details: null